Include calculations of Ne*tau.
Added population-wide Ne*tau, as well as deme-specific Ne*tau. This
is calculated by sampling pairs of infections and seeing if they
coalesce within a certain window of time (currently set to 30 days).
Ne*tau is the inverse of the rate, so opportunity for coalescence
divided by number of coalescent events.